refactor: #169 modified code for search HTTP API
