+++ munetic_express/src/service/search.service.ts
@@ -2,32 +2,9 @@ import { Op } from 'sequelize';
 import * as Status from 'http-status';
 import ErrorResponse from '../modules/errorResponse';
 import { Lesson } from '../models/lesson';
-import { Category } from '../models/category';
-import { User } from '../models/user';
 import * as CategoryService from './category.service';
+import * as UserService from './user.service';
 
-
-export const findCategoryIdByName = async (category_name: string) => {
-    const data = await Category.findOne({
-        where: {
-            name: category_name
-        },
-        attributes: ['id']
-    })
-    if (data === null) return {id: 0};
-    return data;
-};
-
-export const findTutorIdByName = async (tutor_name: string) => {
-    const data = await User.findOne({
-        where: {
-            name: tutor_name
-        },
-        attributes: ['id']
-    })
-    if (data === null) return {id: 0};
-    return data;
-}
 
 export const searchLessonsAll = async () => {
     const data = await Lesson.findAll()
@@ -35,7 +12,7 @@ export const searchLessonsAll = async () => {
 }
 
 export const searchLessonsByCategory = async (category_name: string) => {
-    let result = await findCategoryIdByName(category_name);
+    let result = await CategoryService.findCategoryIdByName(category_name);
     const data = await Lesson.findAll({
         where: {
             category_id: result.id,
@@ -56,7 +33,7 @@ export const searchLessonsByTitle = async (title_name: string) => {
 };
 
 export const searchLessonsByTutor = async (tutor_name: string) => {
-    let result = await findTutorIdByName(tutor_name);
+    let result = await UserService.findTutorIdByName(tutor_name);
     const data = await Lesson.findAll({
         where: {
             tutor_id: result.id,
